enable custom port for test agent
